+++ www/go/base/db/ActiveRecord.php
@@ -3575,12 +3575,20 @@ abstract class ActiveRecord extends \GO\Base\Model{
 			$user_id = GO::user() ? GO::user()->id : 1;
 
 		$acl = new \GO\Base\Model\Acl();
+
 		$acl->usedIn = $this->tableName().'.'.$this->aclField();
 		$acl->ownedBy=$user_id;
 		$acl->entityTypeId = $this->entityType()->getId();
 		$acl->entityId = $this->id;
 		if(!$acl->save()) {
 			throw new \Exception("Could not save ACL: ".var_export($this->getValidationErrors(), true));
+		}
+
+
+		$defaultAclId = static::entityType()->getDefaultAclId();
+		if($defaultAclId) {
+			$defaultAcl = \GO\Base\Model\Acl::model()->findByPk($defaultAclId);
+			$defaultAcl->copyPermissions($acl);
 		}
 
 		$this->{$this->aclField()}=$acl->id;