+++ www/go/core/views/extjs3/license/LicenseDialog.js
@@ -25,7 +25,7 @@ go.license.LicenseDialog = Ext.extend(go.Window, {
 					{
 						xtype: "box",
 						autoEl: "p",
-						html: t('Install your purchased license or try all the features for free and obtain a 60 day trial license from <a target="_blank" class="normal-link" href="https://www.group-office.com">www.group-office.com</a>. Register for an account and get your license now.')
+						html: t('Try the extra features for free and obtain a 60 day trial license from <a target="_blank" class="normal-link" href="https://www.group-office.com">www.group-office.com</a>. Register for an account and get your license now. By purchasing a license you will get support and extra features. Find out more on our website.')
 					},
 					{
 						xtype: "button",